+++ CHANGELOG.md
@@ -11,6 +11,10 @@ vim.keymap.set('i', '<C-z>', '<C-g>u<Esc>[s1z=`]a<C-g>u', { desc = 'Correct late
 
 - FEATURE: Add `tab:> ` to 'listchars' option when `options.extra_ui` is set. This prevents showing `^I` instead of a tab and actual value comes from Neovim's default.
 
+## mini.starter
+
+- FEATURE: Add `sections.pick()` with 'mini.pick' pickers.
+
 ## mini.statusline
 
 - BREAKING FEATURE: Add `search_count` section to default active content.